# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NeverGuess API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          description: Relative path to redirect to after login (must start with `/`).
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
            format: uri
      responses:
        '302':
          description: Redirect to the `returnTo` path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MobileTokenExchangeRequest'
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileTokenExchangeSuccess'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutSuccess'
...
  /audits/{id}/report:
    get:
      tags: [Audits]
      operationId: getReport
      summary: Get the analysis report for an audit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Report returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '401':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found or not ready.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
    MobileTokenExchangeRequest:
      type: object
      required: [code, code_verifier, redirect_uri, state]
      properties:
        code:
          type: string
          minLength: 1
        code_verifier:
          type: string
          minLength: 1
        redirect_uri:
          type: string
          minLength: 1
          format: uri
        state:
          type: string
          minLength: 1
        nonce:
          type: string
          minLength: 1
    MobileTokenExchangeSuccess:
      type: object
      required: [token]
      properties:
        token:
          type: string
    LogoutSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          const: true
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
...
    RiskSeverity:
      type: string
      enum: [critical, high, medium, low]
    RiskEvidence:
      type: object
      description: |
        A concrete pointer into the ingested repo backing a risky assumption.
        `verified` is stamped server-side after cross-checking the cited path
        against the ingested file tree.
      required: [path]
      properties:
        path:
          type: string
        quote:
          type: [string, 'null']
        verified:
          type: boolean
    RiskyAssumption:
      type: object
      required: [title, severity, detail]
      properties:
        title:
          type: string
        severity:
          $ref: '#/components/schemas/RiskSeverity'
        detail:
          type: string
        evidence:
          type: array
          description: Absent on reports generated before citations shipped.
          items:
            $ref: '#/components/schemas/RiskEvidence'
    AffectedArea:
      type: object
      description: |
        Blast radius entry — a file or directory the requested change is
        expected to touch. Paths are model-proposed; `create` entries may
        not exist in the tree yet.
      required: [path, reason, action]
      properties:
        path:
          type: string
        reason:
          type: string
        action:
          type: string
          enum: [modify, create, read]
    AcceptanceCriterion:
      type: object
      required: [title, detail]
      properties:
        title:
          type: string
        detail:
          type: string
    PromptPack:
      type: object
      required: [replit, cursor, copilot, claudeCode, codex]
      properties:
        replit:
          type: string
        cursor:
          type: string
        copilot:
          type: string
        claudeCode:
          type: string
        codex:
          type: string
    Verdict:
      type: string
      enum: [safe, caution, block]
    TestPlan:
      type: object
      required: [vitest, playwright]
      properties:
        vitest:
          type: string
        playwright:
          type: string
    SmokeCheckStatus:
      type: string
      enum: [pass, fail, warn, skip]
    SmokeCheck:
      type: object
      required: [id, label, status, detail]
      properties:
        id:
          type: string
        label:
          type: string
        status:
          $ref: '#/components/schemas/SmokeCheckStatus'
        metric:
          type: string
        detail:
          type: string
    SmokeTestResults:
      type: object
      required: [ranAt, url, skipped, checks]
      properties:
        ranAt:
          type: string
          format: date-time
        url:
          type: string
        skipped:
          type: boolean
        skipReason:
          type: string
        checks:
          type: array
          items:
            $ref: '#/components/schemas/SmokeCheck'
...
    AnalysisUsage:
      type: object
      description: Model response receipt captured from the analysis provider.
      required: [model, requestedModel, reasoningEffort, gateway, generatedAt]
      properties:
        model:
          type: string
        requestedModel:
          type: string
        reasoningEffort:
          type: string
        gateway:
          type: string
          enum: [replit-ai-integrations, openrouter-direct]
        requestId:
          type: [string, 'null']
        promptTokens:
          type: [integer, 'null']
          minimum: 0
        completionTokens:
          type: [integer, 'null']
          minimum: 0
        totalTokens:
          type: [integer, 'null']
          minimum: 0
        reasoningTokens:
          type: [integer, 'null']
          minimum: 0
        generatedAt:
          type: string
          format: date-time
    Report:
      type: object
      required:
        - id
        - auditId
        - source
        - architectureSummary
        - mermaidGraph
        - riskyAssumptions
        - acceptanceCriteria
        - promptPack
        - rolloutNotes
        - riskScore
        - verdict
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        auditId:
          type: string
          format: uuid
        source:
          type: string
        architectureSummary:
          type: string
        mermaidGraph:
          type: string
        riskyAssumptions:
          type: array
          items:
            $ref: '#/components/schemas/RiskyAssumption'
        acceptanceCriteria:
          type: array
          items:
            $ref: '#/components/schemas/AcceptanceCriterion'
        promptPack:
          $ref: '#/components/schemas/PromptPack'
        rolloutNotes:
          type: string
        smokeTestResults:
          oneOf:
            - $ref: '#/components/schemas/SmokeTestResults'
            - type: 'null'
        smokeScreenshotUrl:
          type: [string, 'null']
        analysisUsage:
          oneOf:
            - $ref: '#/components/schemas/AnalysisUsage'
            - type: 'null'
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
        verdict:
          $ref: '#/components/schemas/Verdict'
        testPlan:
          oneOf:
            - $ref: '#/components/schemas/TestPlan'
            - type: 'null'
        affectedAreas:
          oneOf:
            - type: array
              items:
                $ref: '#/components/schemas/AffectedArea'
            - type: 'null'
        shareSlug:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
...